+++ app/Item.php
@@ -248,13 +248,15 @@ class Item extends Model
 
         $data = file_get_contents($file);
 
-        $full = true;
-        if ($new_file = $this->getImagePath($full)) {
-            file_put_contents($new_file, $data);
-            return true;
-        }
+        $this->deleteImage();
 
-        return false;
+        $path = $this->getImagePath($full = true);
+        file_put_contents($path, $data);
+
+        $this->has_image = true;
+        $this->save();
+
+        return true;
     }
 
     public function getOaiUrl()
@@ -705,9 +707,9 @@ class Item extends Model
         return ($this->translate($default_locale)->gallery == 'Slovenská národná galéria, SNG');
     }
 
-    public function scopeHasImage($query)
+    public function scopeHasImage($query, $hasImage = true)
     {
-        return $query->where('has_image', '=', 1);
+        return $query->where('has_image', '=', $hasImage);
     }
 
     public function scopeForReproduction($query)